capstone-project-9900-t16p-undefined-behaviour-main/backend/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==2.6.0
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
